the sync stuff wasn't working, so I removed it; we didn't need/use most of the funcationlity anyways, so I did a refactor; now our code is directly called instead of waiting on Android's ContentResolver - we are the content resolvers!
