+++ openmrs-client/src/main/java/org/openmrs/mobile/dagger/SyncComponent.java
@@ -1,6 +1,5 @@
 package org.openmrs.mobile.dagger;
 
-import org.openmrs.mobile.data.db.impl.PatientIdentifierTypeDbService;
 import org.openmrs.mobile.data.sync.SyncService;
 import org.openmrs.mobile.data.sync.impl.ConceptClassSubscriptionProvider;
 import org.openmrs.mobile.data.sync.impl.DiagnosisConceptSubscriptionProvider;
@@ -13,19 +12,14 @@ import org.openmrs.mobile.data.sync.impl.PersonAttributeTypeSubscriptionProvider
 import org.openmrs.mobile.data.sync.impl.VisitAttributeTypeSubscriptionProvider;
 import org.openmrs.mobile.data.sync.impl.VisitPredefinedTaskSubscriptionProvider;
 import org.openmrs.mobile.data.sync.impl.VisitTypeSubscriptionProvider;
-import org.openmrs.mobile.sync.AndroidSyncService;
-import org.openmrs.mobile.sync.SyncAdapter;
 
 import javax.inject.Singleton;
 
 import dagger.Component;
 
 @Singleton
-@Component(modules = { SyncModule.class, DbModule.class })
+@Component(modules = { DbModule.class })
 public interface SyncComponent {
-	void inject(SyncAdapter syncAdapter);
-	void inject(AndroidSyncService androidSyncService);
-
 	SyncService syncService();
 
 	DiagnosisConceptSubscriptionProvider diagnosisConceptSubscriptionProvider();
