--- openmrs-client/src/main/java/org/openmrs/mobile/net/Authenticator.java
@@ -1,68 +1,0 @@
-package org.openmrs.mobile.net;
-
-import android.accounts.AbstractAccountAuthenticator;
-import android.accounts.Account;
-import android.accounts.AccountAuthenticatorResponse;
-import android.accounts.NetworkErrorException;
-import android.content.Context;
-import android.os.Bundle;
-
-/*
- * Implement AbstractAccountAuthenticator and stub out all of its methods
- * This is a requirement to use syncing though we don't actually use the class
- * https://developer.android.com/training/sync-adapters/creating-authenticator.html
- */
-public class Authenticator extends AbstractAccountAuthenticator {
-
-	// Simple constructor
-	public Authenticator(Context context) {
-		super(context);
-	}
-
-	// Editing properties is not supported
-	@Override
-	public Bundle editProperties(AccountAuthenticatorResponse r, String s) {
-		throw new UnsupportedOperationException();
-	}
-
-	// Don't add additional accounts
-	@Override
-	public Bundle addAccount(AccountAuthenticatorResponse r, String s, String s2, String[] strings, Bundle bundle)
-			throws NetworkErrorException {
-		return null;
-	}
-
-	// Ignore attempts to confirm credentials
-	@Override
-	public Bundle confirmCredentials(AccountAuthenticatorResponse r, Account account, Bundle bundle)
-			throws NetworkErrorException {
-		return null;
-	}
-
-	// Getting an authentication token is not supported
-	@Override
-	public Bundle getAuthToken(AccountAuthenticatorResponse r, Account account, String s, Bundle bundle)
-			throws NetworkErrorException {
-		throw new UnsupportedOperationException();
-	}
-
-	// Getting a label for the auth token is not supported
-	@Override
-	public String getAuthTokenLabel(String s) {
-		throw new UnsupportedOperationException();
-	}
-
-	// Updating user credentials is not supported
-	@Override
-	public Bundle updateCredentials(AccountAuthenticatorResponse r, Account account, String s, Bundle bundle)
-			throws NetworkErrorException {
-		throw new UnsupportedOperationException();
-	}
-
-	// Checking features for the account is not supported
-	@Override
-	public Bundle hasFeatures(AccountAuthenticatorResponse r, Account account, String[] strings)
-			throws NetworkErrorException {
-		throw new UnsupportedOperationException();
-	}
-}
