--- openmrs-client/src/main/java/org/openmrs/mobile/net/impl/AuthenticatorService.java
@@ -1,31 +1,0 @@
-package org.openmrs.mobile.net.impl;
-
-import android.app.Service;
-import android.content.Intent;
-import android.os.IBinder;
-
-import org.openmrs.mobile.net.Authenticator;
-
-/**
- * A bound Service that instantiates the authenticator when started.
- * This is a requirement to use syncing though we don't actually use the class
- * https://developer.android.com/training/sync-adapters/creating-authenticator.html
- */
-public class AuthenticatorService extends Service {
-
-	// Instance field that stores the authenticator object
-	private Authenticator mAuthenticator;
-	@Override
-	public void onCreate() {
-		// Create a new authenticator object
-		mAuthenticator = new Authenticator(this);
-	}
-	/*
-	 * When the system binds to this Service to make the RPC call
-	 * return the authenticator's IBinder.
-	 */
-	@Override
-	public IBinder onBind(Intent intent) {
-		return mAuthenticator.getIBinder();
-	}
-}
