--- openmrs-client/src/main/res/xml/authenticator.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<account-authenticator
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:accountType="org.openmrs.mobile.datasync" />
